requirements.txt
# Python 3 compatible versions - Updated for security
Django==4.2.16
lxml==5.3.0
Scrapy==2.11.2
django-pipeline==3.1.0
psycopg2-binary==2.9.9
w3lib==2.2.1
feedgenerator==2.1.0
brotli==1.1.0
sqlparse==0.5.1
python-dateutil==2.9.0.post0
twisted==24.7.0
